// frontend/src/components/button/enums.ts
export enum TypeButtonEnum {
  Primary = 'primary',
  Secondary = 'secondary',
  Success = 'success',
  Danger = 'danger',
  Warning = 'warning',
  Info = 'info',
  Light = 'light',
  Dark = 'dark',
  Link = 'link',
}

export enum BorderButtonEnum {
  Dashed = 'dashed',
  Default = 'default',
  None = 'none',
}

export enum TextEffectsButtonEnum {
  Italic = 'italic',
  Bold = 'bold',
}

export enum ShapeButtonEnum {
  Circle = 'circle',
  Round = 'round',
  RoundPill = 'rounded-pill',
}

export enum SizeButtonEnum {
  Large = 'large',
  Medium = 'medium',
  Small = 'small',
  Tiny = 'tiny',
}

export enum PositionButtonEnum {
  Left = 'left',
  Right = 'right',
  Center = 'center',
}